Update main page so it points to the new "my work" page
