+++ components/Dashboard/DashboardWrapper.tsx
@@ -28,15 +28,11 @@ const NavLink = ({ href, children }: NavLinkProps) => {
 };
 
 const navigation: { text: string; href: string }[] = [
-  {
-    text: `Allocations`,
-    href: `/`,
-  },
+  { text: 'My work', href: '/' },
   {
     text: `Teams`,
     href: `/teams`,
   },
-  { text: 'My work', href: '/my-work' },
 ];
 
 interface Props {
