+++ cypress/integration/app_config.ts
@@ -4,7 +4,7 @@ describe('app config', () => {
   it('should show the correct value for the workflows app link even after a page navigation event has occurred on the client side', () => {
     cy.visitAs(`/`, AuthRoles.AdminDevGroup);
 
-    cy.get('a').contains('Allocations').click();
+    cy.get('a').contains('My work').click();
 
     cy.location('pathname', { timeout: 60000 }).should('eq', '/');
 
@@ -16,12 +16,6 @@ describe('app config', () => {
 
   it('should show the correct value for the workflows app link even after the browser back button is used to navigate', () => {
     cy.visitAs(`/`, AuthRoles.AdminDevGroup);
-
-    cy.get('a').contains('My work').click();
-
-    cy.location('pathname', { timeout: 60000 }).should('include', '/my-work');
-
-    cy.go('back');
 
     cy.location('pathname', { timeout: 60000 }).should('eq', '/');
